+++ fixes/spacetags.sql
@@ -1,5 +1,5 @@
 BEGIN;
-CREATE OR REPLACE FUNCTION mergeTags(_good bigint, _bad bigint) RETURNS VOID AS
+CREATE OR REPLACE FUNCTION mergeTags(_good INTEGER, _bad INTEGER) RETURNS VOID AS
 $$
 BEGIN
   update things set neighbors = array(select unnest(neighbors) except select _bad union select _good) WHERE neighbors && ARRAY[_good,_bad];
