+++ lua/sql/old.sql
@@ -80,10 +80,10 @@ $$;
 ALTER FUNCTION public.clearneighbors() OWNER TO "user";
 
 --
--- Name: connect(bigint, bigint); Type: FUNCTION; Schema: public; Owner: user
+-- Name: connect(INTEGER, INTEGER); Type: FUNCTION; Schema: public; Owner: user
 --
 
-CREATE FUNCTION connect(_source bigint, _dest bigint) RETURNS boolean
+CREATE FUNCTION connect(_source INTEGER, _dest INTEGER) RETURNS boolean
     LANGUAGE plpgsql
     AS $$
 BEGIN
@@ -96,13 +96,13 @@ END
 $$;
 
 
-ALTER FUNCTION public.connect(_source bigint, _dest bigint) OWNER TO "user";
+ALTER FUNCTION public.connect(_source INTEGER, _dest INTEGER) OWNER TO "user";
 
 --
--- Name: connectmanytoone(bigint[], bigint); Type: FUNCTION; Schema: public; Owner: user
+-- Name: connectmanytoone(INTEGER[], INTEGER); Type: FUNCTION; Schema: public; Owner: user
 --
 
-CREATE FUNCTION connectmanytoone(a bigint[], b bigint) RETURNS void
+CREATE FUNCTION connectmanytoone(a INTEGER[], b INTEGER) RETURNS void
     LANGUAGE plpgsql
     AS $$
 BEGIN
@@ -111,13 +111,13 @@ END;
 $$;
 
 
-ALTER FUNCTION public.connectmanytoone(a bigint[], b bigint) OWNER TO "user";
+ALTER FUNCTION public.connectmanytoone(a INTEGER[], b INTEGER) OWNER TO "user";
 
 --
--- Name: connectone(bigint, bigint); Type: FUNCTION; Schema: public; Owner: user
+-- Name: connectone(INTEGER, INTEGER); Type: FUNCTION; Schema: public; Owner: user
 --
 
-CREATE FUNCTION connectone(a bigint, b bigint) RETURNS void
+CREATE FUNCTION connectone(a INTEGER, b INTEGER) RETURNS void
     LANGUAGE plpgsql
     AS $$
 BEGIN
@@ -126,13 +126,13 @@ END;
 $$;
 
 
-ALTER FUNCTION public.connectone(a bigint, b bigint) OWNER TO "user";
+ALTER FUNCTION public.connectone(a INTEGER, b INTEGER) OWNER TO "user";
 
 --
--- Name: connectonetomany(bigint, bigint[]); Type: FUNCTION; Schema: public; Owner: user
+-- Name: connectonetomany(INTEGER, INTEGER[]); Type: FUNCTION; Schema: public; Owner: user
 --
 
-CREATE FUNCTION connectonetomany(a bigint, b bigint[]) RETURNS void
+CREATE FUNCTION connectonetomany(a INTEGER, b INTEGER[]) RETURNS void
     LANGUAGE plpgsql
     AS $$
 BEGIN
@@ -141,13 +141,13 @@ END;
 $$;
 
 
-ALTER FUNCTION public.connectonetomany(a bigint, b bigint[]) OWNER TO "user";
+ALTER FUNCTION public.connectonetomany(a INTEGER, b INTEGER[]) OWNER TO "user";
 
 --
--- Name: disconnect(bigint, bigint); Type: FUNCTION; Schema: public; Owner: user
+-- Name: disconnect(INTEGER, INTEGER); Type: FUNCTION; Schema: public; Owner: user
 --
 
-CREATE FUNCTION disconnect(a bigint, b bigint) RETURNS void
+CREATE FUNCTION disconnect(a INTEGER, b INTEGER) RETURNS void
     LANGUAGE plpgsql
     AS $$
 BEGIN
@@ -156,17 +156,17 @@ END;
 $$;
 
 
-ALTER FUNCTION public.disconnect(a bigint, b bigint) OWNER TO "user";
+ALTER FUNCTION public.disconnect(a INTEGER, b INTEGER) OWNER TO "user";
 
 --
 -- Name: findtag(text); Type: FUNCTION; Schema: public; Owner: user
 --
 
-CREATE FUNCTION findtag(_name text) RETURNS bigint
+CREATE FUNCTION findtag(_name text) RETURNS INTEGER
     LANGUAGE plpgsql
     AS $$
 DECLARE 
-_id bigint;
+_id INTEGER;
 BEGIN
     LOOP
         SELECT tags.id INTO _id FROM tags WHERE name = _name;
@@ -217,10 +217,10 @@ $$;
 ALTER FUNCTION public.findurisource(_uri text) OWNER TO "user";
 
 --
--- Name: listmedia(bigint[], bigint[], integer, integer); Type: FUNCTION; Schema: public; Owner: user
+-- Name: listmedia(INTEGER[], INTEGER[], integer, integer); Type: FUNCTION; Schema: public; Owner: user
 --
 
-CREATE FUNCTION listmedia(_posi bigint[], _nega bigint[], _offset integer, _limit integer) RETURNS SETOF bigint
+CREATE FUNCTION listmedia(_posi INTEGER[], _nega INTEGER[], _offset integer, _limit integer) RETURNS SETOF INTEGER
     LANGUAGE plpgsql
     AS $$
 DECLARE
@@ -230,7 +230,7 @@ _maxoffset int;
 _oldbottom int;
 _page int DEFAULT 0;
 _base RECORD;
-_derp bigint[];
+_derp INTEGER[];
 BEGIN
     -- get the real offset in media by the 
     _ioffset := _offset;
@@ -287,13 +287,13 @@ END;
 $$;
 
 
-ALTER FUNCTION public.listmedia(_posi bigint[], _nega bigint[], _offset integer, _limit integer) OWNER TO "user";
+ALTER FUNCTION public.listmedia(_posi INTEGER[], _nega INTEGER[], _offset integer, _limit integer) OWNER TO "user";
 
 --
--- Name: mergeadded(bigint, bigint); Type: FUNCTION; Schema: public; Owner: user
+-- Name: mergeadded(INTEGER, INTEGER); Type: FUNCTION; Schema: public; Owner: user
 --
 
-CREATE FUNCTION mergeadded(_a bigint, _b bigint) RETURNS void
+CREATE FUNCTION mergeadded(_a INTEGER, _b INTEGER) RETURNS void
     LANGUAGE plpgsql
     AS $$
 DECLARE
@@ -308,7 +308,7 @@ END
 $$;
 
 
-ALTER FUNCTION public.mergeadded(_a bigint, _b bigint) OWNER TO "user";
+ALTER FUNCTION public.mergeadded(_a INTEGER, _b INTEGER) OWNER TO "user";
 
 --
 -- Name: mergesources(integer, integer); Type: FUNCTION; Schema: public; Owner: user
@@ -360,9 +360,9 @@ CREATE FUNCTION migrateconnections() RETURNS void
     LANGUAGE plpgsql
     AS $$
 DECLARE
-_id bigint;
-_neighbors bigint[];
-_neighbor bigint;
+_id INTEGER;
+_neighbors INTEGER[];
+_neighbor INTEGER;
 _counter int;
 BEGIN
     _counter := 1;
@@ -718,9 +718,9 @@ ALTER TABLE public.comics OWNER TO "user";
 --
 
 CREATE TABLE connections (
-    id bigint NOT NULL,
-    source bigint,
-    dest bigint
+    id INTEGER NOT NULL,
+    source INTEGER,
+    dest INTEGER
 );
 
 
@@ -765,7 +765,7 @@ ALTER TABLE public.derp OWNER TO "user";
 --
 
 CREATE TABLE desktops (
-    id bigint NOT NULL,
+    id INTEGER NOT NULL,
     selected timestamp with time zone DEFAULT clock_timestamp() NOT NULL
 );
 
@@ -778,7 +778,7 @@ ALTER TABLE public.desktops OWNER TO "user";
 
 CREATE TABLE dupes (
     id integer NOT NULL,
-    medium bigint,
+    medium INTEGER,
     hash character varying(28),
     inferior boolean DEFAULT false
 );
@@ -824,7 +824,7 @@ ALTER TABLE public.filesources OWNER TO "user";
 --
 
 CREATE TABLE images (
-    id bigint NOT NULL,
+    id INTEGER NOT NULL,
     animated boolean,
     width integer,
     height integer,
@@ -839,7 +839,7 @@ ALTER TABLE public.images OWNER TO "user";
 --
 
 CREATE TABLE media (
-    id bigint NOT NULL,
+    id INTEGER NOT NULL,
     name text,
     hash character(28),
     created timestamp with time zone,
@@ -919,8 +919,8 @@ ALTER SEQUENCE parsequeue_id_seq OWNED BY parsequeue.id;
 --
 
 CREATE TABLE queryoffsets (
-    posi bigint[],
-    nega bigint[],
+    posi INTEGER[],
+    nega INTEGER[],
     ioffset integer,
     ooffset integer,
     created timestamp with time zone DEFAULT clock_timestamp()
@@ -935,7 +935,7 @@ ALTER TABLE public.queryoffsets OWNER TO "user";
 
 CREATE TABLE randomseen (
     id integer NOT NULL,
-    media bigint,
+    media INTEGER,
     category integer DEFAULT 0
 );
 
@@ -1026,7 +1026,7 @@ ALTER SEQUENCE sources_id_seq OWNED BY sources.id;
 --
 
 CREATE TABLE tags (
-    id bigint NOT NULL,
+    id INTEGER NOT NULL,
     name text
 );
 
@@ -1038,8 +1038,8 @@ ALTER TABLE public.tags OWNER TO "user";
 --
 
 CREATE TABLE things (
-    id bigint NOT NULL,
-    neighbors bigint[]
+    id INTEGER NOT NULL,
+    neighbors INTEGER[]
 );
 
 
@@ -1072,7 +1072,7 @@ ALTER SEQUENCE things_id_seq OWNED BY things.id;
 
 CREATE TABLE uploads (
     uzer integer,
-    media bigint,
+    media INTEGER,
     checked boolean DEFAULT false
 );
 
@@ -1144,7 +1144,7 @@ ALTER SEQUENCE uzer_id_seq OWNED BY uzers.id;
 
 CREATE TABLE uzertags (
     id integer NOT NULL,
-    tag bigint,
+    tag INTEGER,
     uzer integer,
     nega boolean DEFAULT false
 );
@@ -1178,7 +1178,7 @@ ALTER SEQUENCE uzertags2_id_seq OWNED BY uzertags.id;
 --
 
 CREATE TABLE videos (
-    id bigint NOT NULL,
+    id INTEGER NOT NULL,
     width integer,
     height integer,
     fps double precision,
@@ -1197,7 +1197,7 @@ ALTER TABLE public.videos OWNER TO "user";
 CREATE TABLE visited (
     id integer NOT NULL,
     uzer integer,
-    medium bigint,
+    medium INTEGER,
     visits integer DEFAULT 0
 );
 
@@ -1232,7 +1232,7 @@ SET search_path = resultcache, pg_catalog;
 --
 
 CREATE TABLE "q-z8GeQZryx-tM2EYrYYF2bsQFu8" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1244,7 +1244,7 @@ ALTER TABLE resultcache."q-z8GeQZryx-tM2EYrYYF2bsQFu8" OWNER TO "user";
 --
 
 CREATE TABLE "q2AYSoZWiecyi2hcrIMvyYk4bN44" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1258,7 +1258,7 @@ ALTER TABLE resultcache."q2AYSoZWiecyi2hcrIMvyYk4bN44" OWNER TO "user";
 --
 
 CREATE TABLE "q2tu9Yr_bTmwoSvUFp_Kj4P9Zqz4" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1272,7 +1272,7 @@ ALTER TABLE resultcache."q2tu9Yr_bTmwoSvUFp_Kj4P9Zqz4" OWNER TO "user";
 --
 
 CREATE TABLE "q4PGCcCdA1G5R3_mg7B470n7l9CY" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1286,7 +1286,7 @@ ALTER TABLE resultcache."q4PGCcCdA1G5R3_mg7B470n7l9CY" OWNER TO "user";
 --
 
 CREATE TABLE "q79Gdtf4DrdaiBXOGHYgD2nF3jHU" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1298,7 +1298,7 @@ ALTER TABLE resultcache."q79Gdtf4DrdaiBXOGHYgD2nF3jHU" OWNER TO "user";
 --
 
 CREATE TABLE "q8mrNXKvZLjgHvTvOCQSKDdWad8A" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1312,7 +1312,7 @@ ALTER TABLE resultcache."q8mrNXKvZLjgHvTvOCQSKDdWad8A" OWNER TO "user";
 --
 
 CREATE TABLE "qBIHMNcGL1hbwxc9yi3u_JRSjfp4" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1326,7 +1326,7 @@ ALTER TABLE resultcache."qBIHMNcGL1hbwxc9yi3u_JRSjfp4" OWNER TO "user";
 --
 
 CREATE TABLE "qDc8LBI4UllkApWBYlu0wt_bC1LQ" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1340,7 +1340,7 @@ ALTER TABLE resultcache."qDc8LBI4UllkApWBYlu0wt_bC1LQ" OWNER TO "user";
 --
 
 CREATE TABLE "qECwqVfHpq6jp3RIqaaJY2ZePXdI" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1352,7 +1352,7 @@ ALTER TABLE resultcache."qECwqVfHpq6jp3RIqaaJY2ZePXdI" OWNER TO "user";
 --
 
 CREATE TABLE "qImfoIkXHIwyAodZKw4kK4tyFA1w" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1364,7 +1364,7 @@ ALTER TABLE resultcache."qImfoIkXHIwyAodZKw4kK4tyFA1w" OWNER TO "user";
 --
 
 CREATE TABLE "qJ4YjoQrU-xLlODIqlgk7Tw_mh9U" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1376,7 +1376,7 @@ ALTER TABLE resultcache."qJ4YjoQrU-xLlODIqlgk7Tw_mh9U" OWNER TO "user";
 --
 
 CREATE TABLE "qJScke_npFNV1IyuqA6B3yy1tf54" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1388,7 +1388,7 @@ ALTER TABLE resultcache."qJScke_npFNV1IyuqA6B3yy1tf54" OWNER TO "user";
 --
 
 CREATE TABLE "qJ_8yPV2cHI_w3cfmjd-Tqw3w7jo" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1400,7 +1400,7 @@ ALTER TABLE resultcache."qJ_8yPV2cHI_w3cfmjd-Tqw3w7jo" OWNER TO "user";
 --
 
 CREATE TABLE "qLjXVNHr-6Vu7pfBLWxUVD7eWRik" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1414,7 +1414,7 @@ ALTER TABLE resultcache."qLjXVNHr-6Vu7pfBLWxUVD7eWRik" OWNER TO "user";
 --
 
 CREATE TABLE "qMn2vW6eu8zFmZMs4GpMKJ1QxXl0" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1428,7 +1428,7 @@ ALTER TABLE resultcache."qMn2vW6eu8zFmZMs4GpMKJ1QxXl0" OWNER TO "user";
 --
 
 CREATE TABLE "qNyP-Y09e8L_hrpCsOGra9hf33MY" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1440,7 +1440,7 @@ ALTER TABLE resultcache."qNyP-Y09e8L_hrpCsOGra9hf33MY" OWNER TO "user";
 --
 
 CREATE TABLE "qO5c54xbyLU96WV-tuhzW8WtrMpk" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1454,7 +1454,7 @@ ALTER TABLE resultcache."qO5c54xbyLU96WV-tuhzW8WtrMpk" OWNER TO "user";
 --
 
 CREATE TABLE "qOBOgBg1tdm_wEKZOhnCEy8pOBnY" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1468,7 +1468,7 @@ ALTER TABLE resultcache."qOBOgBg1tdm_wEKZOhnCEy8pOBnY" OWNER TO "user";
 --
 
 CREATE TABLE "qOyxm4RUyqHLwB2o2EERmkVddWOA" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1480,7 +1480,7 @@ ALTER TABLE resultcache."qOyxm4RUyqHLwB2o2EERmkVddWOA" OWNER TO "user";
 --
 
 CREATE TABLE "qOzmy0CzzXUOd821CtcBxXeGjD_g" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1492,7 +1492,7 @@ ALTER TABLE resultcache."qOzmy0CzzXUOd821CtcBxXeGjD_g" OWNER TO "user";
 --
 
 CREATE TABLE "qRNvBW_WPZnzLIN8kc94AjkYjQ_o" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1506,7 +1506,7 @@ ALTER TABLE resultcache."qRNvBW_WPZnzLIN8kc94AjkYjQ_o" OWNER TO "user";
 --
 
 CREATE TABLE "qS95_81u_yUfuCwgIWuqiyTQrUSw" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1518,7 +1518,7 @@ ALTER TABLE resultcache."qS95_81u_yUfuCwgIWuqiyTQrUSw" OWNER TO "user";
 --
 
 CREATE TABLE "qWWT2HVTBvKmiEC9IOxLC9qOmTUk" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1530,7 +1530,7 @@ ALTER TABLE resultcache."qWWT2HVTBvKmiEC9IOxLC9qOmTUk" OWNER TO "user";
 --
 
 CREATE TABLE "qXA5dnqFqedLxJcUcV1O7jNtUWEo" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1542,7 +1542,7 @@ ALTER TABLE resultcache."qXA5dnqFqedLxJcUcV1O7jNtUWEo" OWNER TO "user";
 --
 
 CREATE TABLE "qXjXFlRs6aVlZOvg3JkfbQVGoBA8" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1554,7 +1554,7 @@ ALTER TABLE resultcache."qXjXFlRs6aVlZOvg3JkfbQVGoBA8" OWNER TO "user";
 --
 
 CREATE TABLE "q_SitDDhgmMWnXQLgAYfORJxEako" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1566,7 +1566,7 @@ ALTER TABLE resultcache."q_SitDDhgmMWnXQLgAYfORJxEako" OWNER TO "user";
 --
 
 CREATE TABLE "qiJTqYrhoApmGp5F7_b2GWT_IWQY" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1578,7 +1578,7 @@ ALTER TABLE resultcache."qiJTqYrhoApmGp5F7_b2GWT_IWQY" OWNER TO "user";
 --
 
 CREATE TABLE "qjnDdR9eOeon3iJagRuC-AEj_MHQ" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1592,7 +1592,7 @@ ALTER TABLE resultcache."qjnDdR9eOeon3iJagRuC-AEj_MHQ" OWNER TO "user";
 --
 
 CREATE TABLE "qky6gfav1QCvK-uob2JViml6lq0U" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1606,7 +1606,7 @@ ALTER TABLE resultcache."qky6gfav1QCvK-uob2JViml6lq0U" OWNER TO "user";
 --
 
 CREATE TABLE "qn9tadg88rgQkpjdEA-tPmBCBiUQ" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1620,7 +1620,7 @@ ALTER TABLE resultcache."qn9tadg88rgQkpjdEA-tPmBCBiUQ" OWNER TO "user";
 --
 
 CREATE TABLE "qoG8bFGXsZ5N0vCY2GFcKgzfpdgQ" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1634,7 +1634,7 @@ ALTER TABLE resultcache."qoG8bFGXsZ5N0vCY2GFcKgzfpdgQ" OWNER TO "user";
 --
 
 CREATE TABLE "qoRERGH1Waygq-wgtbvB6LukZ19Q" (
-    id bigint,
+    id INTEGER,
     name text
 );
 
@@ -1646,7 +1646,7 @@ ALTER TABLE resultcache."qoRERGH1Waygq-wgtbvB6LukZ19Q" OWNER TO "user";
 --
 
 CREATE TABLE "qsDHHa_fvhUKA11Vnn0wTOWnbozo" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
@@ -1694,7 +1694,7 @@ ALTER SEQUENCE queries_id_seq OWNED BY queries.id;
 --
 
 CREATE TABLE "qvzHILWpLpuwn96H9Cc_l0qaOaBU" (
-    id bigint,
+    id INTEGER,
     name text,
     type text,
     "array" text[]
